backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn==23.0.0
